Tune error reporting when PATH_SYMEX_FORK is enabled
When PATH_SYMEX_FORK, we have effectively a tree of
processes with the usual terminology for "parent",
"sibling" and "child". If a child finds a bug, the
parent should not report any other bugs (if any).

We achieve this by having the parent check the exit
status code of its child processes. The rationale is
that this is more efficient than POSIX inter-process
communication. Of course, this does not rule out the
possibility that siblings report their bugs. But this
is a reasonable trade-off with the low complexity
of the design and its performance characteristics.

We use WNOHANG | WNOWAIT to "poll" the exit status
code of children (if any) in a nonblocking way. And
our interpretation of the POSIX standard concerning
multiple calls to waitpid() is illustrated by the
following C program:

#include <stdio.h>
#include <errno.h>
#include <stdlib.h>
#include <unistd.h>
#include <assert.h>
#include <sys/wait.h>

int main()
{
  pid_t pid = fork();

  if (pid < 0)
  {
      fprintf(stderr, "Failed to fork\n");
      return EXIT_FAILURE;
  }
  else if (pid == 0)
  {
    // child process
    _exit(5);
  }
  else
  {
    // parent process
    int status;
    pid_t child_pid;

    // here it doesn't matter that we're
    // racing with child's _exit(5)
    child_pid = waitpid(pid, &status, 0);
    assert(child_pid == pid);

    if (errno == ECHILD)
    {
      fprintf(stderr, "Got ECHILD (bad)\n");
    }

    // wait again 
    child_pid = waitpid(pid, &status, 0);
    assert(child_pid == -1);

    if (errno == ECHILD)
    {
      fprintf(stdout, "Got ECHILD (good)\n");
    }

    assert(WIFEXITED(status));
    assert(WEXITSTATUS(status) == 5);
  }

  return EXIT_SUCCESS;
}



git-svn-id: http://www.cprover.org/svn/cbmc/trunk@3641 6afb6bc1-c8e4-404c-8f48-9ae832c5b171