Terminate correctly
Made the domain correclty find the fixed point. The merge operation
returns whether the merge actually changed any values by implementing a
operator== for each of the abstract objects.

Further, when merging a map, if a key is absent from one map then we
remove it from the other unless the map is bottom.