Basic implementation of the sensitive pointer
Created class that tracks an abstract object that it is pointing to.
Involved modifying constructors to take exprt rather than constant_exprt
since would want to create the pointer using an address_of_exprt which
isn't a constant_exprt. I'm not sure if this is a good change but
couldn't think of an alternative.