Fixed crash bug in writing to dereference of pointer to pointer
Fixed a number of issues with pointer to pointer that were out by one
(dereference).

Specifically, when we are writing to a dereference, the pointer itself
remains unchanged, so what we return from the write dereference is just
a clone of te pointer.

When reading, we don't eval the address_of exprt as don't know how to
handle this, instead we eval the object of the address of (since we are
dereferencing to call the read_dereference function anyway).