Bug fixes for the array abstract object
Enabling array tests
Handler in eval for array since we don't need to run any transforms on
it, we can just create the object directly.
Corrected base class constructor to not return bottom when an unknown
write is performed.
Fixed issue where if we didn't find an entry we tried to return it.