Corrected implementation of flow sensitivity
It used to assume(!guard) when next==to under the assumption that this
means we are not taking the jump. This doesn't work if the GOTO is
jumping to the next line.

Instead we first check that the target of the goto statement is not the
same as the next line. If it is, we don't assume anything since any
condition could be both true or false on the next line.

If this isn't the case, then we use the target of the jump and the to
variable to determine whether the jump was taken or not.