Correcting implemenations of the write methods for pointer and struct
When writing to a dereferenced pointer when we have the two value
(top/bottom) domain, if the pointer is top, it stays as top and if it is
bottom, it stays as bottom. However, in all other cases it becomes top.

Similarly for structs, the struct should be top unless it is bottom.