Fixing constant propogation test
The variable_sensitivity_domaint::ai_simplify was retuning true if it
simplied, but it should have been returning true if no simplification
was performed (as specified in ai.h).

The tests also had the wrong number of goto statements, there are two in
the original program, both of which are removed

Some of the tests had unescaped strings so I fixed them and made the more
consistent with the other tests as enabling them.

Some tests were disabled, they look like the should work but currently return
the wrong number of simplified assignments. Created issue
diffblue/cbmc-toyota#112 to look into this more detail.