Adding reasonable behaviour for opqaue function calls
When we have a function call without a body, we could set everything to
top (this would be more correct). To keep consistent, we default at the
moment to the more reasonable, less sound behaviour of only topping
pointers passed in and the return value.

Updated the test description to now be able to validate the simple
assert.