Implementing a write_stack to represent writing to a addressable location
Deals with storing a stack to write to things like:

&a[1]
&s.comp
&a[x] + 1

By building a stack as parsing the expression tree.

It does not support pointer arithmetic outside the scope of an array
(e.g. within a struct)

This will be used to store pointer values.

Disabled unit tests since don't have the prerequistes in this branch