[DO NOT MERGE] Track which expressions give rise to data dependencies
This is a customer special feature which tracks which expressions are the
source of any data dependencies, displaying those expressions as part of the
output. JSON output also includes the 'certainty' level of the dependency.
At the moment, this is a hard coded/enabled feature. It should be made
controlled by a command line option before it gets considered for merging, but
PR'ing anyway for review comments.